Migrate stabilized correlations to `boilercv`
