+++ docs/conf.py
@@ -7,12 +7,12 @@ from pathlib import Path
 from ruamel.yaml import YAML
 from sphinx.application import Sphinx
 
+from boilercv.correlations.dimensionless_bubble_diameter import get_equations
 from boilercv_docs import DOCS, PYPROJECT, chdir_docs
 from boilercv_docs.intersphinx import get_ispx, get_rtd, get_url
 from boilercv_docs.nbs import init_nb_env
 from boilercv_docs.patch_nbs import patch_nbs
 from boilercv_docs.types import IspxMappingValue
-from boilercv_pipeline.correlations.dimensionless_bubble_diameter import get_equations
 
 # ! Initialization
 patch_nbs()
