--- pipeline/boilercv_pipeline/correlations/dimensionless_bubble_diameter/equations.py
@@ -1,1 +1,0 @@
-"""Equations."""
