+++ pyproject.toml
@@ -13,10 +13,12 @@ classifiers = ["License :: OSI Approved :: MIT License"]
 dependencies = [
     "matplotlib>=3.7.2",
     "numpy>=1.24.4",
+    "numpydantic>=1.0.0",
     "pandas[hdf5,performance]>=2.0.2",
     "pillow>=10.0.0",
-    "pydantic>=2.7.1",                     # ? RootModel improvements
+    "pydantic>=2.7.1",
     "pytz>=2023.3",
+    "sympy>=1.12",
     "xarray[accel,io,parallel]>=2023.7.0",
 ]
 [project.optional-dependencies]
@@ -42,9 +44,11 @@ source = [
 ]
 [tool.coverage.report]
 ignore_errors = true
+
 [tool.fawltydeps]
 deps = ["pyproject.toml"]
 code = ["src"]
+ignore_undeclared = ["_typeshed"]
 
 [tool.pytest.ini_options]
 addopts = '''
