+++ src/boilercv/correlations/dimensionless_bubble_diameter/types.py
