+++ src/boilercv/correlations/types.py
@@ -1,8 +1,7 @@
-"""Static type annotations used in {mod}`boilercv_pipeline`."""
+"""Static type annotations used in {mod}`boilercv.correlations`."""
 
 from collections.abc import Callable, Mapping
 from dataclasses import asdict, dataclass
-from functools import wraps
 from pathlib import Path
 from types import SimpleNamespace
 from typing import (
@@ -37,22 +36,25 @@ from boilercv.morphs.contexts import ContextValue
 if TYPE_CHECKING:
     from _typeshed import DataclassInstance
 
+# * MARK: Type aliases
+
+LiteralKeys: TypeAlias = _LiteralGenericAlias
+"""Keys."""
 # ? Notebook handling
 NbProcess: TypeAlias = Callable[[Path, SimpleNamespace], None]
 """Notebook process."""
 Stage: TypeAlias = Literal["large_sources", "sources", "filled"]
 """Stage."""
-
 # ? `numpy` array shapes
 AnyShape: TypeAlias = Shape["*"]  # noqa: F722
 """Any shape."""
 Vector: TypeAlias = Shape["*"]  # noqa: F722
 """Vector shape."""
-
 # ? Equations
 Expectation: TypeAlias = float | NDArray[Vector, float]  # pyright: ignore[reportInvalidTypeArguments]
 """Expectation."""
 
+# * MARK: Type variables
 
 P = TypeVar("P", contravariant=True)
 """Contravariant type to represent parameters."""
@@ -69,92 +71,14 @@ CV = TypeVar("CV", bound=ContextValue, contravariant=True)
 CVL = TypeVar("CVL", bound="DataclassInstance | Mapping[Any, Any]", contravariant=True)
 """Context value-like type."""
 
+# * MARK: Protocols
+
 
 class Transform(Protocol[P, Ps, R]):  # noqa: D101
     def __call__(self, i: P, /, *args: Ps.args, **kwds: Ps.kwargs) -> R: ...  # noqa: D102
 
 
-def StrSerializer(  # noqa: N802; Can't inherit from frozen
-    when_used: Literal["always", "unless-none", "json", "json-unless-none"] = "always",
-) -> WrapSerializer:
-    """Serialize as string."""
-    return WrapSerializer(_str, when_used=when_used)
-
-
-def _str(v: Any, nxt: SerializerFunctionWrapHandler):
-    """Stringify or try next serializer."""
-    try:
-        return str(v)
-    except ValueError:
-        return nxt(v)
-
-
-@overload
-def TypeValidator(typ: type[K], mode: Literal["before"]) -> BeforeValidator: ...
-@overload
-def TypeValidator(typ: type[K], mode: Literal["wrap"]) -> WrapValidator: ...
-@overload
-def TypeValidator(typ: type[K], mode: Literal["after"]) -> AfterValidator: ...
-@overload
-def TypeValidator(typ: type[K], mode: Literal["plain"]) -> PlainValidator: ...
-@overload
-def TypeValidator(typ: type[K]) -> PlainValidator: ...
-
-
-# * MARK: `TypeVar`s and `TypeAlias`es for annotations
-
-
-Validator: TypeAlias = Literal["before", "wrap", "after", "plain"]
-"""Validators."""
-
-
-VALIDATORS: dict[Validator, Any] = {
-    "before": BeforeValidator,
-    "wrap": WrapValidator,
-    "after": AfterValidator,
-    "plain": PlainValidator,
-}
-"""Validators."""
-
-
-def TypeValidator(typ: type[K], mode: Validator = "plain") -> Any:  # noqa: N802; Can't inherit from frozen
-    """Validate type."""
-
-    def validate(v: K) -> K:
-        if isinstance(v, typ):
-            return v
-        raise ValueError(f"Input should be a valid {typ}")
-
-    return VALIDATORS[mode](validate)
-
-
-LiteralKeys: TypeAlias = _LiteralGenericAlias
-"""Keys."""
-
-
-# * MARK: Annotation parts
-
-
-def contextualized(context_value_type: type[CVL]):
-    """Contextualize function on `context_value_type`."""
-
-    def contextualizer(f: Transform[P, Ps, R]) -> Callable[[P, CVL], R]:
-        @wraps(f)
-        def unpack_kwds(v: P, context_value: CVL) -> R:
-            if not isinstance(context_value, context_value_type):
-                raise ValueError(  # noqa: TRY004 so Pydantic catches it
-                    f"Expected context value type '{context_value_type}', got '{type(context_value)}."
-                )
-            context_mapping = (
-                context_value
-                if isinstance(context_value, Mapping)
-                else asdict(context_value)
-            )
-            return f(v, **context_mapping)  # pyright: ignore[reportCallIssue]
-
-        return unpack_kwds
-
-    return contextualizer
+# * MARK: Validator
 
 
 def validator(context_value_type: type[CV]):
@@ -187,6 +111,73 @@ def validator(context_value_type: type[CV]):
     return validator_maker
 
 
+Kind: TypeAlias = Literal["latex", "sympy"]
+"""Kind."""
+Equation: TypeAlias = Literal[
+    "florschuetz_chao_1965", "isenberg_sideman_1970", "akiyama_1973", "yuan_et_al_2009"
+]
+"""Equation."""
+
+# * MARK: Serializers
+
+
+def StrSerializer(  # noqa: N802; Can't inherit from frozen
+    when_used: Literal["always", "unless-none", "json", "json-unless-none"] = "always",
+) -> WrapSerializer:
+    """Serialize as string."""
+    return WrapSerializer(_str, when_used=when_used)
+
+
+def _str(v: Any, nxt: SerializerFunctionWrapHandler):
+    """Stringify or try next serializer."""
+    try:
+        return str(v)
+    except ValueError:
+        return nxt(v)
+
+
+# * MARK: Validators
+
+
+Validator: TypeAlias = Literal["before", "wrap", "after", "plain"]
+"""Validators."""
+
+
+@overload
+def TypeValidator(typ: type[K], mode: Literal["before"]) -> BeforeValidator: ...
+@overload
+def TypeValidator(typ: type[K], mode: Literal["wrap"]) -> WrapValidator: ...
+@overload
+def TypeValidator(typ: type[K], mode: Literal["after"]) -> AfterValidator: ...
+@overload
+def TypeValidator(typ: type[K], mode: Literal["plain"]) -> PlainValidator: ...
+@overload
+def TypeValidator(typ: type[K]) -> PlainValidator: ...
+def TypeValidator(typ: type[K], mode: Validator = "plain") -> Any:  # noqa: N802; Can't inherit from frozen
+    """Validate type."""
+
+    def validate(v: K) -> K:
+        if isinstance(v, typ):
+            return v
+        raise ValueError(f"Input should be a valid {typ}")
+
+    return VALIDATORS[mode](validate)
+
+
+VALIDATORS: dict[Validator, Any] = {
+    "before": BeforeValidator,
+    "wrap": WrapValidator,
+    "after": AfterValidator,
+    "plain": PlainValidator,
+}
+"""Validators."""
+
+S = TypeVar(
+    "S", bound=Annotated[LiteralKeys, TypeValidator(str), StrSerializer("json")] | str
+)
+"""Keys representing symbols."""
+
+
 @dataclass
 class SympifyParams(ContextValue):
     """Sympify parameters."""
@@ -203,14 +194,17 @@ def sympify_expr(expr: str | sympy.Eq, sympify_params: SympifyParams):
     return sympify(expr, **asdict(sympify_params))
 
 
-trivial = sympy.Eq(1, 0, evaluate=False)
-"""Trivial equation that won't evaluate to a `True` instance of `sympy.boolean.linalg.BooleanAtom`."""
+# * MARK: Annotated types for Pydantic
+
 
 Basic: TypeAlias = Annotated[
     sympy.Basic, TypeValidator(sympy.Basic), StrSerializer("json")
 ]
 """Annotated {class}`~sympy.core.basic.Basic` suitable for use in Pydantic models."""
 
+
+trivial = sympy.Eq(1, 0, evaluate=False)
+"""Trivial equation that won't evaluate to a `True` instance of `sympy.boolean.linalg.BooleanAtom`."""
 
 Eq: TypeAlias = Annotated[
     Basic,
@@ -224,12 +218,13 @@ Eq: TypeAlias = Annotated[
     ),
 ]
 """{data}`~boilercv_pipeline.types.Basic` narrowed to {class}`~sympy.core.relational.Eq`."""
-# * MARK: Annotations
 
-
-Expr = Annotated[
+Expr: TypeAlias = Annotated[
     Basic,
     BeforeValidator(validator(SympifyParams)(sympify_expr)),
     TypeValidator(sympy.Expr, "after"),
 ]
 """{data}`~boilercv_pipeline.types.Basic` narrowed to {class}`~sympy.core.relational.Eq`."""
+
+EQ = TypeVar("EQ", bound=Eq | str)
+"""SymPy symbolic equation or string type."""
