+++ tests/boilercv_tests/test_correlations/__init__.py
@@ -1,0 +1,1 @@
+"""Correlation tests."""
