--- tests/boilercv_tests/test_equations/__init__.py
@@ -1,1 +1,0 @@
-"""Equations tests."""
